portfolio-risk-app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
yfinance==0.2.33
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1

# Authentication & Security
pyotp==2.9.0
qrcode[pil]==7.4.2
phonenumbers==8.13.26
twilio==8.10.0

# Email & Communication
aiosmtplib==3.0.1
jinja2==3.1.2
premailer==3.10.0

# Rate Limiting & Caching
slowapi==0.1.9
python-multipart==0.0.6

# Utilities
python-magic==0.4.27
pillow==10.1.0